Renamed folders. Refactored the csv method. Fixed an error using the color attribute.
